+++ program/src/lib.rs
@@ -1,5 +1,3 @@
-#![feature(destructuring_assignment)]
-
 #[macro_use]
 pub mod error;