+++ changelog.md
@@ -1,5 +1,10 @@
 # Mango Program Change Log
 
+## v3.2.2
+Deployed:
+1. Get rid of destructuring assignment feature
+2. Use impact bid/ask for calculating funding (100 contracts)
+
 ## v3.2.1
 Deployed: Nov 1, 2021 at 18:09:05 UTC | Slot: 104,689,370
 1. If perp market is added before spot market, fix decimals to 6